added test C682
