+++ cypress/support/fragments/eholdings/eHoldingsProviders.js
@@ -2,6 +2,7 @@ import {
   Accordion,
   Button,
   ListItem,
+  PaneHeader,
   RadioButton,
   Section,
   Spinner,
@@ -76,4 +77,8 @@ export default {
       packagesSection.find(ListItem({ text: including(filterStatuses.notSelected) })).absent(),
     ]);
   },
+
+  verifyProviderHeaderTitle: title => {
+    cy.expect(PaneHeader(title).exists());
+  }
 };
